ci: run push build only on master
semantic-release was failing to push tags to publish because pushing tags required this build to
pass. We will no longer require tag pushes to have a sucessfulbuild. Furthermore, to prevent double
builds on pull requests, this only run on pushes to the master branch (as well as any pull request).

fixes #86 after-the-fact