fix(package.json): add directory to package.json within monorepo
Previosly, when RenovateBot would update this package in other projects, it was unable to read the
CHANGELOG.md file. This is because the package lives in a monorepo, so RenovateBot didn't know where
to look for the changelog of the specific package being updated. This follows RenovateBot's
suggestion, and should fix the issue. See the suggestion here:
https://github.com/renovatebot/renovate/issues/2926#issuecomment-575094237